- up to 5.0.8888, add desktop file
Changed files:
    teamviewer.spec -> 1.3

+++ teamviewer.spec
@@ -1,16 +1,19 @@
 # TODO
-# - use system wine
-# - desktop file
+# - use system wine (bundles unmodified wine 1.1.41)
+%define		buildid	8888
+%define		rel		0.1
 Summary:	TeamViewer Remote Control Application
 Name:		teamviewer
 Version:	5.0
-Release:	0.3
+Release:	%{buildid}.%{rel}
 License:	Proprietary; includes substantial Free Software components, notably the Wine Project.
 Group:		Applications/Networking
 Source0:	http://www.teamviewer.com/download/%{name}_linux.tar.gz
-# NoSource0-md5:	2ff6ec6410f61b8f5d96d2057d00f886
+# NoSource0-md5:	10ba96fd81ac520f66c0f52cf70836a0
 NoSource:	0
 Source1:	%{name}.sh
+Source2:	%{name}.desktop
+Source3:	%{name}.png
 ExclusiveArch:	%{ix86}
 BuildRoot:	%{tmpdir}/%{name}-%{version}-root-%(id -u -n)
 
@@ -31,20 +34,30 @@ buy a license for commercial use, visit the webpage.
 %setup -q -n %{name}5
 install -p %{SOURCE1} teamviewer
 
+ver=$(strings ".wine/drive_c/Program Files/TeamViewer/Version5/TeamViewer.exe" | grep %{version}.%{buildid})
+if [ -z "$ver" ]; then
+	exit 1
+fi
+
 %install
 rm -rf $RPM_BUILD_ROOT
-install -d $RPM_BUILD_ROOT{%{_appdir},%{_bindir}}
+install -d $RPM_BUILD_ROOT{%{_appdir},%{_bindir},%{_desktopdir},%{_pixmapsdir}}
 cp -a .wine teamviewer $RPM_BUILD_ROOT%{_appdir}
 ln -s %{_appdir}/teamviewer $RPM_BUILD_ROOT%{_bindir}
+cp -a %{SOURCE2} $RPM_BUILD_ROOT%{_desktopdir}
+cp -a %{SOURCE3} $RPM_BUILD_ROOT%{_pixmapsdir}
 
 %clean
 rm -rf $RPM_BUILD_ROOT
 
 %files
 %defattr(644,root,root,755)
-%doc license_teamviewer_en.txt copyrights_en.txt license_foss.txt
-%doc %lang(de) copyrights_de.txt linux_faq_de.txt
+%doc license_foss.txt
+%doc %lang(de) *_DE.txt
+%doc %lang(en) *_EN.txt
 %attr(755,root,root) %{_bindir}/teamviewer
+%{_desktopdir}/%{name}.desktop
+%{_pixmapsdir}/%{name}.png
 %dir %{_appdir}
 %attr(755,root,root) %{_appdir}/teamviewer
 # XXX: temp & ugly, until system wine works
